move approveAll method to the UserProjectService
